+++ CHANGELOG.md
@@ -487,6 +487,9 @@ Miscellaneous:
   using the "Store settings" button in the widget settings panel, available by
   clicking the window icon right next to the widget title.
 
+- Neuron navigator: neuron widget nodes (like the active neuron display) show
+  now also the initial creator of the neuron and the initial creation time.
+
 - Auto-completion should now be faster by displaying only 15 filtering results.
 
 - Volume widget: box volumes can now be created also based on the active